requirements.txt
Django>=1.8,<1.9
djangorestframework
psycopg2
ipython
coolname
